Create a base image modeled after ci-core (#1)
Create a base image that all images defined in the repository will inherit from in some form.

This image described in this PR is the cardboardci/ci-core image, with some minor modifications to support the new monolithic approach to dockerfiles. The images that inherit from the base are not required to work off `HEAD`, but that is not a constraint, just a technical implementation detail.

The aim of having the base image here, is it allows all of the images to clearly see the list of available dependencies, scripts, and elements that the base image has.