Enable github actions on the  (#2)
Enable GitHub Actions for the base container image, resolving bugs with namespaces and `git` version.

This PR moves the workflow for building the base container image into the `workflows` folder, and resolves some namespace bugs introduced with the merged of the `versions` and `vars` steps. 

Additionally as an update for git has rolled out, I have upgraded to the latest version of git.