Scripts for automatic upgrading of dependency lists (#5)
Scripts for retrieving the latest version of a dependency from the relevant package manager. This removes the need for manual lookup, and allows the process to be completed from a central makefile.

The usage style is intended to be `make <package>-<image>`, and will attempt to get the latest available version for that dependency.

For now, the dependencies of images will be list in txt-style files, that will be read in and installed on the fly. Long term is for each of the dependencies to be described in a yml or json format.