Generic icon that is used for the container images (#6)
Add in the generic icon that is used to represent container images, that is a placeholder for the time being.

All of the containers have an icon associated with them, so that they can be looked up in a visual manner. The ones used at the moment are placeholders, as they will later be replaced with images relevant to the internals of the container.