gitlab - Deploy README to docker hub description (#57)
When the README of the image changes, it should update the readme on Docker Hub. This allows for common gitlab commands to be listed alongside the image itself.

I expect to later revise this to have a common descriptor template for every image, that references back to the central website. The advantage to that model is a better 'How to use' this image model, rather than attempting to link to existing yml files in source (for CI providers).