The luarocks dependency check is hanging on an interactive check (#87)
The luarocks downloader setup was encountering an interactive blocker, which prevented the check from progressing. This PR ensures that these will run in  a non-interactive execution.

Additionally the luarocks check was occurring in the latex image, when it should have been occurring in the luacheck image.

Setting up a downloader image would be better, as that would allow for testing to ensure scenarios like this don't happen in the future.