Inline auto-commit to allow adjustment to downloader (#88)
Inline the auto-commit action as a defined action within the repository. Replacing the usages in all the version updater workflows.

I expect to inline some core actions into the repository for the time being while they are unstable. When they become stable, they can be moved back into the `actions` repository for the organization.