Upgrade all images to the monolith base image (#93)
Switch from the original base image `ci-core` to that of `base`, which is defined in this repository.

A `digest` argument has been introduced to all dockerfiles to simplify the upgrade process. Now this line can be updated using a script, or altered for the purposes of debugging.

Additionally this upgrades all the base versions to the latest, as some of them were a bit out of date.