Merge all version workflows into a single job (#165)
Merge all version workflows into a single workflow responsible for version bumps.

This is done in preparation for consolidating the dependency checking code. I want to shift this to have a better more 'dependabot' feel to the upgrading of each dependency. At the moment any kind of bumping work is all-or-nothing for an image.

This changes starts the ground-work for a more complex version bumping system.

Additional work will need to be done to `auto-commit` to support a more templated description of the dependency (see dependabot as an example).