Use GitHub Container Registry to replace the packages usage (#169)
Switch to using the new `ghcr.io` for hosting the container images.

This is one of the final steps in replacing DockerHub with GitHub for the serving point of the container images. Secrets management is a bit of an issue in that it requires a github PCAT, but hopefully this will be resolved shortly to allow for use of the runners github token.

The packages will be manually cleaned up as part of this work.