Support building awscli image directly from bazel (#171)
Support building the generated dockerfile from bazel for the AWSCLI image.

This does leave the option of removing the `Dockerfile` from the source, and just generating it on the fly. For now I'm going to keep both of them running while I explore a better package management system. Ideally the package management system than the `provision/*list` system.

Limitations of rules_docker with respect to windows prevent it from being used in this case. I am fine with that for now, as it gives a good opportunity to fiddle around with various bazel features like runfiles & genrules.