Image prototype for using rules_docker with images (#172)
A prototype awscli image built with rules_docker instead of using dockerfiles.

The current bazel rules are based around generating a dockerfile then using a genrule to run `docker build`. With rules_docker this no longer relies on the generation of a dockerfile. 

To move this forward into the rest of the images, this will need to prove out a CI/CD strategy, wrappers to reduce redeclaration and a means of upgrade automation through buildozer.

This is now a viable option as I have established a code-server environment based on linux, which removes the issues that this previously had with no windows compatibility.