Configure the CI/CD pipeline for a bazel type build process (#174)
Configure the rules_docker bazel prototype for running in github actions.

This is part of the work to verify that the same workflow done for Dockerfiles can be done with bazel. The images need to be able to be built from github actions, and deployed to the github container registry. Testing is confirmed to work as expected, and modifications will need to be made to support all types of version dependencies.

Dependency upgrading seems to be the primary blocker to leveraging rules_docker in the codebase as a whole.