Support devcontainers for build on windows (#183)
Support windows development with devcontainers, allowing for containerized development.

The rules_docker bazel rules don't support windows that well, but devcontainers allows the containerization of the development environment. As long as the the remote containers extension is installed in visual studio, this allows for opening the repository in a bazel container.

The docker socket is mounted so the downloading of images will work with the host docker daemon.