feat(awscli): Support bazel style builds for awscli (#204)
Support bazel style builds for the AWSCLI.

This isn't complete support for the bazel style builds, but is intended to just a starter for modelling the other images.

The intention is to replace all images with bazel, using `run_and_commit` to fill in any gaps. Then leverage codify the gaps into bazel rules.