feat(awscli): Migrate from Dockerfile to bazel (#205)
Migrate the awscli image from using Dockerfile to bazel for construction.

This switches from using Dockerfile with hadolint for static analysis, to construction being handled entirely through bazel. Test scaffolding is handled by bazel, running the container-structure-tests under the hood.

AWSCli was a simple image to begin with, and will probably be deprecated now that AWS has an official image. However this is a good test image to start with for switching over images from using Dockerfiles to using bazel.