build(lint): Remove hadolint from build pipeline (#209)
Remove hadolint from the build pipeline in preparation for the bazel shift.

Each of the images has its own hadolint configuration file, even if unused.  Stripping these out as one makes it easier to mold each of the images over to the bazel format without requiring every migration to handle the removal process. Hadolint won't be needed for the final bazel image with the new way of constructing images.

Additionally the `.dockerignore` file has been removed, as the Dockerfile will soon be deprecated.