feat(bazel): Configure bazel starters for all images (#210)
Configure bazel starter environments for all of the images.

All of the placeholders are modelled after the `awscli` image, as its a basic image with only a single installed dependency. The labels have been copied over to the `image` target. All tests and components of these images are just placeholders for the time being.

These placeholders have been added to reduce the number of moving parts associated with moving all the images from Dockerfiles to bazel. As most of the images are going to have similar targets, this provides something that each of the images can  modify until it fits the expected model.