feat(svgtools): Use bazel for builds instead of dockerfiles (#217)
Use bazel to build the svgtools image, removing the dockerfile build process.

This converts from using Dockerfile to build the image, to using bazel to construct it. The old build sources are removed, and replaced with just the bazel driven build process. Equivalents for static analysis are not needed during the transition period, as it does not use shell commands directly to build the images.