feat(pylint): Convert to bazel for builds (#218)
Convert the pylint image to use bazel instead of dockerfiles.

This converts to using bazel to build the dockerfile, and switches from the cardboardci base to using ubuntu. This does mean that some commands will be missing from the compiled image. For now, this should be fine as switching to bazel is intended to streamline this as a whole.

This uses the ubuntu image as a base due to the current base image being incompatible with pip installations.