feat(hugo): Convert the hugo image to use bazel (#226)
Convert the hugo image to use bazel for builds instead of dockerfiles.

This converts to using bazel to build the image. There is room for optimization (in terms of `deb` downloads) but for now this is sufficient for getting started with the build pipeline of an image.

Additionally the old dockerfile pipeline has been cleaned up.