remove(latex): Deprecate the latex image (#240)
Deprecate the latex image and remove it from source.

This removes the latex image from the list of dockerfiles, and the associated latex image will remain on github registry for now, until a better option has been decided for building latex docs (or just use HTML to PDF).

Packaging and redistributing the entire latex ecosystem doesn't make sense for builds of this nature. It makes more sense to package together a common set as necessary (e.g. policy doc, essay, resume, etc) that contains all of the common useful elements.