feat(pdf2htmlex): Switch image to use bazel builds (#227)
Switch the build process to work entirely through bazel.

The compilation process for fontforge and pdf2htmlEX is done inside the image for ease-of-conversion. Ideally these should be compiled to an output artifact that can be integrated into the main image without requiring the additional build dependencies or artifacts.

Additionally the dockerfile pipeline artifacts have been cleaned up.