chore(dockerfiles): Cleanup dockerfile elements (#246)
Cleanup the dockerfile elements used for building images.

Images are no longer built with dockerfiles, so it is no longer necessary to support any of the build tooling and scripting that worked with that old model. New systems will need to be built to replace some of theses (automatic version bumps), but for now this is dead-code.

Workflows have been renamed to make file discovery easier when working with different types of workflows (see versions.yml).