docs(repository): Migrate icon to the repository metadata (#247)
Migrate the icon of the repository over to the repository metadata convention.

The convention for my repositories is intended to make the sourcing of icons and metadata about a repository easier. This is done so that automated categorization and organizing can occur with all of my repositories, without coupling any implementation to a specific service.