feat(images): Use in-repo base image for builds (#249)
Convert images to use in-repo base image, rather than prebuilt external.

This helps make the development process easier as it no longer relies on two different changes when bumping the base image. Any change to the base image can be used as a means to trigger an update on all of the images, and confirm they work as expected.