feat(website): Inline website repository into dockerfiles (#252)
Inline the website repository into the project, removing the old website repository.

This inlines the documentation of the website repository into the project, creating the website on the domain `cardboardci.jrbeverly.me`. The site is hosted by GitHub Actions.

Additionally this will simplify the process of listing tags and other components when constructing the site, as the identifiers for the images will be available in-source.