feat(images): Wrap repetitive commit_and_run targets (#254)
Convert repetitive commit_and_run targets to macros wrapping intent.

This wraps repetitive commit_and_run targets into macros that clarifies the dependencies of the image with a `download_and_install` style. Eventually it would be ideal for these macros to be converted to macros that perform the `download` and `install` similar to the `apt` rules.

Additionally the entire codebase has been formatted as per buildifier guidelines. Future work will be done to add in a buildifier lint check as a requirement for any BUILD/WORKSPACE changes.