feat(tests): Centralize metadata and user-level tests (#255)
Require all images to meet the same standard of metadata and user-level tests.

This imposes the same container-struct-tests onto all of the images, ensuring a certain level of metadata is met for the images. Additionally all images are expected to be restricted to the user/group `cardboardci:cardboardci` during execution.

Bazel formatting tests have been added as part of the commit, and all images have been formatted to meet buildifier expectations.

Sudo access is not granted at this time, as installation of dependencies is not expected at runtime.