fix(upgrader): Use docker pull for digest (#258)
Use docker pull to determine the latest image digest available for the ubuntu base image.

This switches to using `docker pull` to determine the latest base image, as the registry API wasn't returning the `focal` image in the top 100 latest images. The API doesn't support a query mechanism, and using bash to perform the iteration seems like it would be error-prone.

Alternative tools like `docker-browse` can be explored for looking up the latest image without requiring a pull.