build(upgrader): Polish base image upgrade workflow (#260)
Resolve issues that were occurring with the ubuntu base image upgrade workflow. 

This cleans up a series of issues that were happening with the upgrader script, polishing how this is done and ensuring that it is more stable at triggering workflows multiple times in the event of a missed pull request merge. Previously this workflow could fail if too many tags were returned, or if branches were not merged in a timely manner.

Additionally a workaround has been applied for buildozer due to limitations with how edits to WORKSPACE files occur.