fix(upgrader): Wrong digest pulled from docker pull (#270)
The digest used from docker pull/inspect is incorrect.

Pull the digest from the registry V2 api, and get the latest available tag from the v1 api. This will ensure that the correct digest is passed on to the WORKSPACE file, rather than the digest of the manifest itself.

The selection of the latest tag has been split out to its own action to reduce the complexity in the digest action.