docs(wrappers): Document design for integrity checks (#274)
Document the design for integrity checks on downloaded packages.

The bazel rules at present do not verify the integrity of the downloaded dependencies at any step in the process. This is a design for including a mechanism for integrity verification of each of the dependencies. The download process as a single step to avoid potential problems that may be encountered with the underlying package managers.

Integrity is split into its own rule so that downloads do not fail in the event of a bad integrity check.