docs(images): Source image pages from image directory (#278)
Source the contents of the image webpages from the images directory.

This copies the README sources for each of the images over to the website, ensuring that the documentation for each image remains up to date. The theme of the website has been changed to better reflect the project, with updated icons.

Additionally update the project icon while revising the website documentation.