docs(images): Standard template for all images (#283)
Configure a standard template for all of the images.

This setups a common way of viewing each of the images on the cardboardci webpage, and understanding how to interact with them. The README doesn't need to contain any common test samples or use cases, as that can be baked into the images labels, then understood by the caller.