docs(conventions): Initial drafts of image intentions (#284)
Drafts of the initial intentions behind the images of CardboardCI.

Include a series of recommendations and general points on how images can be designed as a whole. The intention at this time is not completeness of the documentation, but a starting set that can have smaller modifications made as the desire arises.