fix(docs|tests): Fixes to images to match doc expectations (#289)
Resolve issues where the documentation was not referencing the correct version command.

The documentation has been updated to use the correct version commands for each of the images. Any images that were not in an exactly working state have been updated with a version check command to ensure that they are installed as expected.